+++ scripts/fastqtsvfile.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python
 
 import glob
-import itertools
 import csv
 import os
-import xmltodict
-import functools
 import argparse
+from build_meta import *
+
 
 def tsvbuild(path, pattern, list_of_paths, tsv_name):
     """builds a tsv file from a directory of paired files
@@ -53,125 +52,20 @@ def tsvbuild(path, pattern, list_of_paths, tsv_name):
     return tsv_name
 
 
-def build_metadata(identifier, path, list_of_paths, input_dict={}):
-    """Generates metadata from xml file
-
-    Args:
-        identifier: string, Name of the dictionary information will be extracted from
-        path: string, location of xml file
-        input_dict: optional variable, dictionary, contains metadata information
-    
-    Returns:
-        list of strings, metadata information
-    Notes:
-        'DESIGN/SAMPLE_DESCRIPTOR/PRIMARY_ID',
-        'DESIGN/LIBRARY_DESCRIPTOR/LIBRARY_NAME,
-        'PLATFORM/ILLUMINA/INSTRUMENTAL_MODEL', ILLUMINA, SC
-       
-    """
-    # retrieve parent directory from path and specify experiment xml file
-    exper_dict = input_dict
-    exp_loc = os.path.abspath(os.path.join(path, os.pardir))
-    exp_id = os.path.split(exp_loc)[1]
-    exp_name = exp_loc + os.sep + exp_id + '.experiment.xml'
-    try:
-        experiment = next(gen_dict_extract(identifier, exper_dict[exp_name]), None)
-    except KeyError:
-        exper_dict[exp_name] = xmldictconv(exp_name)
-        experiment = next(gen_dict_extract(identifier, exper_dict[exp_name]), None)
-     return (exper_dict, dictquery(experiment, list_of_paths))
-
-
-def xmldictconv(xml_name):
-    """Convert xml fie to an ordered dictionary
-
-    Args:
-        xml_name: string, path to xml file
-
-    Returns:
-        Dictionary version of xml file or empty dictionary if file not found
-    """
-    try:
-        with open(xml_name, 'r') as xml_file:
-            output_dict = xmltodict.parse(xml_file.read())
-        return output_dict
-    except OSError:
-        return {}
-
-
-def grouper(iterable, size, fillvalue=None):
-    """Fetches elements from an iterable at the specified size at a time
-
-    Args:
-        iterable: traversable object
-        size: Integer, specifies size of slice
-        fillvalue: Integer or String symbol to fill missing chunk space.
-
-    Returns:
-        tuple, chunk of an iterable of specified size
-    """
-    chunk = [iter(iterable)] * size
-    return itertools.zip_longest(*chunk, fillvalue=fillvalue)
-
-
-def gen_dict_extract(value, var, des_dict=None):
-    """Search a list of nested dictionaries
-
-    Args:
-        value: Object being search
-        var: iterable containing dictionaries or lists
-        des_dict: default dictionary
-
-    Yields:
-        Dictionary containing desired values
-    """
-    # not optimal
-    try:
-        for k, v in var.items():
-            if v == value:
-                yield des_dict
-    except AttributeError:
-        pass
-    else:
-        # try excepts all the way down?
-        if isinstance(v, dict):
-            for result in gen_dict_extract(value, v, des_dict):
-                yield result
-        elif isinstance(v, list):
-            for d in v:
-                des_dict = d
-                for result in gen_dict_extract(value, d, des_dict):
-                    yield result
-
-
-def dictquery(input_dict, listofpath):
-    """Fetches data from nested dictionary
-    Args:
-        input_dict: nested Dictionary
-        listofpath: path of desired value
-    
-    Returns:
-        list of values at path specified by input list
-    """
-    output_list = []
-    for path in listofpath:
-        output_list.append(functools.reduce(dict.get, path.split('/'), input_dict))
-    return output_list
-
 # path, pattern, list_of_paths, tsv_name)
 
 def parse_args():
     """Parses arguments
     Args:
-    
     Returns:
         arguments
     """
     parser = argparse.ArgumentParser()
-    parser.add_argument('path', type=string, help='location of parent folder')
-    parser.add_argument('pattern', type=string, help='location of parent folder')
+    parser.add_argument('path', help='location of parent folder')
+    parser.add_argument('pattern', help='location of parent folder')
     parser.add_argument('lists_of_paths', type=list, help='location of parent folder')
-    parser.add_argument('tsv_name', type=string, help='location of parent folder')
+    parser.add_argument('tsv_name', help='location of parent folder')
+    args = parser.parse_args()
     return (args.path, args.pattern, args.list_of_paths, args.tsv_name)
